impliment update and delete goals for authorized user

+++ backend/controllers/goalController.js
@@ -1,15 +1,12 @@
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const asyncHandler = require("express-async-handler");
 
-
-
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({user: req.user.id});
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
-
-
 
 const setGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
@@ -32,6 +29,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found!");
   }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body);
   res.status(200).json(updatedGoal);
 });
@@ -42,6 +53,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found!");
+  }
+  
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized");
   }
 
   await goal.deleteOne();
